requirements.txt
absl-py==0.14.1
airsim==1.6.0
cachetools==4.2.4
# certifi==2021.5.30
certifi==2022.6.15
charset-normalizer==2.0.6
cloudpickle==2.0.0
cycler==0.10.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
gym==0.21.0
idna==3.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
numpy==1.21.2
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
# stable-baselines3==1.2.0
stable-baselines3==1.6.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
# torch==1.9.1
torch==1.12.1
tornado==4.5.3
typing-extensions==3.10.0.2
# urllib3==1.26.7
urllib3==1.26.12
Werkzeug==2.0.2
wincertstore==0.2

sympy==1.11
scipy==1.9.1
selenium==4.4.3
tqdm==4.64.1
wandb==0.13.4
openpyxl==3.0.10
xlsxwriter == 3.0.3

jdatetime == 4.1.0
# TensorFlow == 2.10.0